% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransLasso.R
\name{agg.fun}
\alias{agg.fun}
\title{agg.fun}
\usage{
agg.fun(B, X.test, y.test, total.step = 10, selection = F)
}
\arguments{
\item{total.step}{(default: 10)}

\item{selection}{(default: F)}
}
\description{
agg.fun
}
\examples{
... show how to create vars
agg.fun(B, X.test, y.test)
}
