% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransLasso.R
\name{las.kA}
\alias{las.kA}
\title{Trans-Lasso Oracle algorithm method}
\usage{
las.kA(X, y, A0, n.vec, lam.const = NULL, l1 = T)
}
\arguments{
\item{X}{is the training data set}

\item{y}{is the target value set}

\item{A0}{is the informative auxiliary set}

\item{n.vec}{PLACEHOLDER}

\item{lam.const}{is a tuning parameter (default: NULL)}

\item{l1}{is a boolean indicating l1-sparse characterization of contrast vectors (T) or l0 (F) (default: T)}
}
\value{
a named list containing beta.kA, w.kA, and lam.const
}
\description{
Trans-Lasso Oracle algorithm method
}
