% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransLasso.R
\name{mse.fun}
\alias{mse.fun}
\title{Computing the mean squared error}
\usage{
mse.fun(beta, est, X.test = NULL)
}
\arguments{
\item{beta}{is the vector to compare against}

\item{est}{is the estimate vector}

\item{X.test}{is the test training data set (default: NULL)}
}
\value{
is a named list containing est.err and pred.err
}
\description{
Computing the mean squared error
}
