% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransLasso.R
\name{plot.translasso}
\alias{plot.translasso}
\title{Plot translasso object}
\usage{
\method{plot}{translasso}(obj, ...)
}
\arguments{
\item{obj}{is the translasso object to be plotted}

\item{...}{are any extra parameters}
}
\description{
Plot translasso object
}
