% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransLasso.R
\name{print.translasso}
\alias{print.translasso}
\title{Print translasso object}
\usage{
\method{print}{translasso}(x, ...)
}
\arguments{
\item{x}{is the translasso object to be printed}

\item{...}{are any extra parameters}
}
\description{
Print translasso object
}
