% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransLasso.R
\name{Trans.lasso}
\alias{Trans.lasso}
\title{Trans Lasso algorithm method}
\usage{
Trans.lasso(X, y, n.vec, I.til, l1 = T)
}
\arguments{
\item{X}{is the training data set}

\item{y}{is the target value set}

\item{n.vec}{PLACEHOLDER}

\item{I.til}{is a subset of n.vec containing roughly half its elements}

\item{l1}{is a boolean indicating l1-sparse characterization of contrast vectors (T) or l0 (F) (default: T)}
}
\value{
A named list containing beta.hat, theta.hat, rank.pi, beta.pool, and theta.pool
}
\description{
Trans Lasso algorithm method
}
