% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translasso-package.R
\docType{package}
\name{translasso-package}
\alias{translasso}
\alias{translasso-package}
\title{translasso: Perform Transfer Learning with Lasso Models}
\description{
Fits a two-stage Lasso model where information is transferred from a source dataset to a target dataset. This method estimates an initial model using the combined datasets, computes an offset based on the source data, and then fits a second Lasso model on the target dataset adjusted for this offset.
}
\author{
\strong{Maintainer}: Aaron Molstad \email{amolstad@umn.edu}

Authors:
\itemize{
  \item Ting-Chien Huang \email{tchuang1290@gmail.com}
}

}
\keyword{internal}
