% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransLasso.R
\name{agg.fun}
\alias{agg.fun}
\title{agg.fun}
\usage{
agg.fun(B, X.test, y.test, total.step = 10, selection = F)
}
\arguments{
\item{B}{PLACEHOLDER}

\item{X.test}{PLACEHOLDER}

\item{y.test}{PLACEHOLDER}

\item{total.step}{PLACEHOLDER (default: 10)}

\item{selection}{PLACEHOLDER (default: F)}
}
\value{
is a named list containing theta, beta, and beta.ew
}
\description{
agg.fun
}
