% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransLasso.R
\name{las.kA}
\alias{las.kA}
\title{Trans-Lasso Oracle}
\usage{
las.kA(X, y, A0, n.vec, lam.const = NULL, l1 = T)
}
\arguments{
\item{X}{is the training data}

\item{y}{is the target value set}

\item{A0}{is the informative set}

\item{n.vec}{PLACEHOLDER}

\item{lam.const}{PLACEHOLDER (default: NULL)}

\item{l1}{is a boolean indicating l1-sparse characterization of contrast vectors (default: T)}
}
\value{

}
\description{
Trans-Lasso Oracle
}
