% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransLasso.R
\name{las.kA}
\alias{las.kA}
\title{Trans-Lasso Oracle}
\usage{
las.kA(X, y, A0, n.vec, lam.const = NULL, l1 = T)
}
\arguments{
\item{X}{PLACEHOLDER}

\item{y}{PLACEHOLDER}

\item{A0}{PLACEHOLDER}

\item{n.vec}{PLACEHOLDER}

\item{lam.const}{PLACEHOLDER (default: NULL)}

\item{l1}{PLACEHOLDER (default: T)}
}
\description{
Trans-Lasso Oracle
}
