% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransLasso.R
\name{las.kA}
\alias{las.kA}
\title{Trans-Lasso Oracle}
\usage{
las.kA(X, y, A0, n.vec, lam.const = NULL, l1 = T)
}
\arguments{
\item{lam.const}{(default: NULL)}

\item{l1}{(default: T)}
}
\description{
Trans-Lasso Oracle
}
\examples{
... show how to set vars
las.kA(X, y, A0, n.vec)
}
