% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransLasso.R
\name{mse.fun}
\alias{mse.fun}
\title{Computing the MSE}
\usage{
mse.fun(beta, est, X.test = NULL)
}
\arguments{
\item{beta}{PLACEHOLDER}

\item{est}{PLACEHOLDER}

\item{X.test}{PLACEHOLDER (default: NULL)}
}
\description{
Computing the MSE
}
