% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransLasso.R
\name{summary.translasso}
\alias{summary.translasso}
\title{Summarize translasso object}
\usage{
\method{summary}{translasso}(object, ...)
}
\arguments{
\item{...}{are any extra parameters}

\item{x}{is the translasso object to be summarized}
}
\description{
Summarize translasso object
}
