% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransLasso.R
\name{Trans.lasso}
\alias{Trans.lasso}
\title{Trans Lasso algorithm method}
\usage{
Trans.lasso(X, y, n.vec, I.til, l1 = T)
}
\arguments{
\item{X}{PLACEHOLDER}

\item{y}{PLACEHOLDER}

\item{n.vec}{PLACEHOLDER}

\item{I.til}{PLACEHOLDER}

\item{l1}{PLACEHOLDER (default: T)}
}
\value{
A something
}
\description{
Trans Lasso algorithm method
}
