% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransLasso.R
\name{Trans.lasso}
\alias{Trans.lasso}
\title{Trans Lasso main method}
\usage{
Trans.lasso(X, y, n.vec, I.til, l1 = T)
}
\arguments{
\item{l1}{(default: T)}
}
\value{
A something
}
\description{
Trans Lasso main method
}
\examples{
... how to setup vars
Trans.lasso(X, y, n.vec, I.til)
}
