% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransLasso.R
\name{Trans.lasso.sp}
\alias{Trans.lasso.sp}
\title{A method for comparison: Trans-Lasso(l1). It has the same pipeline of Trans.lasso() but with sparsity index R_k=|w^{(k)}-β|_1 and a naive aggregation (empirical risk minimization)}
\usage{
Trans.lasso.sp(X, y, n.vec, I.til, l1 = T)
}
\arguments{
\item{X}{PLACEHOLDER}

\item{y}{PLACEHOLDER}

\item{n.vec}{PLACEHOLDER}

\item{I.til}{PLACEHOLDER}

\item{l1}{PLACEHOLDER (default: T)}
}
\description{
A method for comparison: Trans-Lasso(l1). It has the same pipeline of Trans.lasso() but with sparsity index R_k=|w^{(k)}-β|_1 and a naive aggregation (empirical risk minimization)
}
